# Make a specialized `elAttr` function