Since this (and `el`) is such a common case, it should speed up DOM
building.

~2x speedup on static DOM building benchmark (which uses only `elAttr`).

Before:

```
+----------------+----------+-------+-------+
| Name           | Op/s     | % max | +-(%) |
+----------------+----------+-------+-------+
| js 10          | 16359.82 | 100   | 4.50  |
| js_c 10        | 13945.26 | 85.24 | 3.05  |
| js_m 10        | 13272.51 | 81.13 | 3.42  |
| static mono 10 | 2718.16  | 16.61 | 1.39  |
| static 10      | 2715.30  | 16.60 | 1.61  |
+----------------+----------+-------+-------+
```

After:

```
+----------------+----------+-------+-------+
| Name           | Op/s     | % max | +-(%) |
+----------------+----------+-------+-------+
| js 10          | 17785.71 | 100   | 4.50  |
| js_c 10        | 14925.13 | 83.92 | 2.62  |
| js_m 10        | 14440.06 | 81.19 | 2.53  |
| static mono 10 | 6103.31  | 34.32 | 1.75  |
| static 10      | 6186.16  | 34.78 | 1.37  |
+----------------+----------+-------+-------+
```